rewrite pauli part
